# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TOHFA Platform API
  version: 1.0.0
  summary: Golden-thread API for the TOHFA farmer / customer / admin platform.
  description: |
    # TOHFA Platform API

    This document is **the contract**. Every client — the Farmer mobile app, the Customer
    mobile app, the Admin web/mobile app, and every internal service SDK — is generated
    from this file. Nothing that is not described here is part of the v1 surface. If a
    client needs a field, it is added here first, reviewed, and only then implemented.

    The "golden thread" is the end-to-end path that carries value through the platform:

    `farmer onboarding -> certification verification -> fair price ceiling -> produce
    listing -> admin approval / counter-offer -> purchase order -> goods receipt +
    quality check -> inventory batch -> channel allocation -> farm-anonymous catalog ->
    cart -> wallet-first checkout -> order -> warehouse fulfilment + OTP -> invoice ->
    farmer payout`

    Every endpoint below is annotated with two vendor extensions:

    * `x-permission` — the permission code from `docs/rbac.json` that the
      gateway checks before the handler runs. `public` means the endpoint is
      unauthenticated (it also carries `security: []`).
    * `x-business-rules` — the client-locked business rules the handler enforces
      server-side. Rules are never enforced only in the client.

    ## Business rules

...

...
    rules that no v1 endpoint enforces yet.

...

...

...
  contact:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
  /auth/register/customer:
    post:
      tags: [Auth]
      operationId: registerCustomer
      summary: Self-register as a customer
      description: |
        Customers self-register and are activated instantly once their mobile OTP is
        verified — no admin review. Farmers cannot use this endpoint; they submit an
        application (see Farmer Onboarding).
      security: []
      x-permission: public
      x-business-rules: [BR-32]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CustomerRegistrationRequest' }
            example:
              mobile: '+919876543210'
              fullName: Priya Raman
              email: priya@example.com
              password: 'C0rrect-Horse-9'
              preferredLocale: en
              preferredWarehouseId: 3f1b7c2e-2a44-4f0e-9b1a-6b0d2f0f1a01
      responses:
        '201':
          description: Registration accepted; an OTP has been dispatched to the mobile number.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegistrationAccepted' }
              example:
                userId: 8c0f0f6a-6f1a-4f1e-8d5a-2b7c9e0a1122
                status: PENDING_OTP
                otpExpiresAt: '2026-08-18T06:35:00Z'
                resendAvailableAt: '2026-08-18T06:31:00Z'
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/ServerError' }
  /auth/otp/send:
    post:
      tags: [Auth]
      operationId: sendOtp
      summary: Send a 6-digit OTP
      description: 6-digit code, 60-second resend timer, 3-attempt lockout (BR-32).
      security: []
      x-permission: public
      x-business-rules: [BR-32]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OtpSendRequest' }
            example: { mobile: '+919876543210', purpose: LOGIN }
      responses:
        '202':
          description: OTP dispatched.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OtpChallenge' }
              example:
                challengeId: 0b2d0f2e-8b1c-4c1a-9f0a-9d3e6c1b2a30
                expiresAt: '2026-08-18T06:35:00Z'
                resendAvailableAt: '2026-08-18T06:31:00Z'
                attemptsRemaining: 3
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/ServerError' }
  /auth/otp/verify:
    post:
      tags: [Auth]
      operationId: verifyOtp
      summary: Verify an OTP and mint a token pair
      description: |
        Exchanges the challenge id and the 6-digit code for an access/refresh token pair. A
        wrong code decrements `attemptsRemaining`; the third failure locks the challenge and a
        new one must be requested (BR-32).

        Optionally carries a BR-39 `linkToken` (from `POST /auth/oauth/{provider}`'s
        NOT_LINKED response) to complete an OAuth link atomically, in the same transaction as
        this OTP verification: an existing mobile number gets the identity linked to that
        account; a brand-new number gets the account created (straight to ACTIVE) and then
        linked. The response then also carries `oauthProfile` so the client can prefill Step 1
        of registration.
      security: []
      x-permission: public
      x-business-rules: [BR-32, BR-39]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OtpVerifyRequest' }
            example:
              challengeId: 0b2d0f2e-8b1c-4c1a-9f0a-9d3e6c1b2a30
              code: '482915'
      responses:
        '200':
          description: OTP accepted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OtpVerifyResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/ServerError' }
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Log in with mobile + password
      description: |
        An account may hold multiple admin roles. When more than one role is active the
        response sets `requiresRoleSelection` and lists `availableRoles`; the client shows
        the multi-role selector and repeats the call with `roleCode`.
      security: []
      x-permission: public
      x-business-rules: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
            example:
              mobile: '+919876543210'
              password: 'C0rrect-Horse-9'
              deviceId: 6f2a-android-pixel7
              platform: android
      responses:
        '200':
          description: Authenticated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
              example:
                accessToken: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI4YzBmMGY2YSJ9.sig
                refreshToken: 9c1d0b7a4f5e4a3c8d2b1e0f6a7b8c9d
                tokenType: Bearer
                expiresIn: 900
                requiresRoleSelection: false
                user:
                  id: 8c0f0f6a-6f1a-4f1e-8d5a-2b7c9e0a1122
                  fullName: Priya Raman
                  userType: customer
                  roles: []
                  preferredLocale: en
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/ServerError' }
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      summary: Exchange a refresh token for a new token pair
      description: |
        Rotates the pair: the presented refresh token is invalidated as the new one is issued,
        so a replayed refresh token is treated as theft and revokes the whole session family.
      security: []
      x-permission: public
      x-business-rules: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: New token pair issued; the presented refresh token is rotated out.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/ServerError' }
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Revoke the current session
      description: |
        Revokes the refresh token behind the current session. Access tokens already issued stay
        valid until they expire, so clients must also drop their in-memory copy.
      x-permission: auth.session.revoke_own
      x-business-rules: []
      responses:
        '204': { description: Session revoked. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
  /auth/forgot-password:
    post:
      tags: [Auth]
      operationId: forgotPassword
      summary: Start a password reset via mobile OTP
      description: |
        Sends a password-reset OTP to the registered mobile number. The response is identical
        whether or not the number is registered, so this cannot be used to enumerate accounts
        (BR-32).
      security: []
      x-permission: public
      x-business-rules: [BR-32]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mobile]
              properties:
                mobile: { $ref: '#/components/schemas/Mobile' }
      responses:
        '202':
          description: If the mobile number exists, an OTP has been dispatched. Always 202 to avoid account enumeration.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OtpChallenge' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/ServerError' }
  /auth/reset-password:
    post:
      tags: [Auth]
      operationId: resetPassword
      summary: Complete a password reset
      description: |
        Sets a new password once the reset OTP has been verified, and terminates every other
        active session for that account (BR-32).
      security: []
      x-permission: public
      x-business-rules: [BR-32]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challengeId, code, newPassword]
              properties:
                challengeId: { type: string, format: uuid }
                code: { type: string, pattern: '^[0-9]{6}$' }
                newPassword: { type: string, minLength: 10, maxLength: 128 }
      responses:
        '204': { description: Password reset; all existing sessions are revoked. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/ServerError' }
  /auth/sessions/{id}/terminate:
    post:
      tags: [Auth]
      operationId: terminateSession
      summary: Terminate another admin's session
      description: Super Admin only. Terminating your own session is a logout, not this call.
      x-permission: auth.session.terminate_other
      x-business-rules: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      responses:
        '204': { description: Session terminated. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: Current principal, roles and effective permissions
      description: |
        Returns the authenticated principal with its roles, warehouse or zone scope and
        effective permission codes — the payload every client shell reads to decide what to
        render.
      x-permission: auth.principal.view_own
      x-business-rules: []
      responses:
        '200':
          description: The authenticated principal.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUser' }
              example:
                id: 2a9f0d1b-7c3e-4a55-9f0b-1e2d3c4b5a60
                fullName: Anand Kumar
                mobile: '+919812345678'
                userType: admin
                preferredLocale: ta
                roles:
                  - code: SUB_WH_ADMIN
                    warehouseId: 3f1b7c2e-2a44-4f0e-9b1a-6b0d2f0f1a01
                    warehouseName: Coonoor
                permissions: [goods_receipt.create, order.verify_otp, wallet.cash_topup]
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/ServerError' }
  /auth/oauth/{provider}:
    post:
      tags: [Auth]
      operationId: loginWithOAuth
      summary: Log in with a verified Google/Facebook identity (BR-39)
      description: |
        OAuth is an alternate login method for an already mobile-verified account, or a
        data-prefill convenience during first-time registration — it is NEVER a way to create
        or activate an account without mobile+OTP (BR-39).

        The server verifies `token` against the provider (Google ID token signature/audience,
        or a Facebook Graph API `/me` call) and looks it up by `(provider, providerSubjectId)`:

        * **Linked already** — logs in exactly like `/auth/login`, including
          `requiresRoleSelection` when the account holds more than one role.
        * **Not linked** — 200 `NOT_LINKED` with whatever profile fields the provider gave
          (`fullName`, `email`, `photoUrl` — none guaranteed) and a `linkToken`: a short-lived
          (10 minute), single-use, server-signed token. No account is created and no tokens are
          issued. The client runs the normal `/auth/otp/send` + `/auth/otp/verify` flow and
          passes `linkToken` to `/auth/otp/verify` to complete the link.
      security: []
      x-permission: public
      x-business-rules: [BR-39]
      parameters:
        - name: provider
          in: path
          required: true
          schema: { $ref: '#/components/schemas/OAuthProvider' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OAuthLoginRequest' }
            example:
              token: eyJhbGciOiJSUzI1NiIsImtpZCI6IjazEyMyJ9.eyJzdWIiOiIxMDc2NTQzMjEwIn0.sig
              deviceId: 6f2a-android-pixel7
              platform: android
      responses:
        '200':
          description: Either an authenticated session (identity already linked) or a NOT_LINKED profile + linkToken.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/TokenPair'
                  - $ref: '#/components/schemas/OAuthNotLinkedResponse'
              example:
                status: NOT_LINKED
                profile:
                  fullName: Priya Raman
                  email: priya@example.com
                  photoUrl: https://lh3.googleusercontent.com/a/photo.jpg
                linkToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ0eXAiOiJvYXV0aF9saW5rIn0.sig
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/ServerError' }
  /auth/me/oauth/link:
    post:
      tags: [Auth]
      operationId: linkOwnOAuthIdentity
      summary: Link a Google/Facebook identity to your own account (BR-39)
      description: |
        Self-service linking for an already-authenticated caller: verifies `token` against the
        named provider and attaches the resulting identity to the caller's own account.
        Rejects with `OAUTH_IDENTITY_ALREADY_LINKED` if that identity is already linked to a
        DIFFERENT account. Linking an identity already linked to the caller's own account is a
        no-op that returns the existing link.
      x-permission: auth.oauth.link_own
      x-business-rules: [BR-39]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OAuthLinkRequest' }
            example: { provider: google, token: eyJhbGciOiJSUzI1NiJ9.eyJzdWIiOiIxMDc2NTQzMjEwIn0.sig }
      responses:
        '200':
          description: Identity linked (or already linked to this same account).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OAuthIdentity' }
              example:
                provider: google
                email: priya@example.com
                linkedAt: '2026-09-17T06:35:00Z'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/ServerError' }
  /auth/me/oauth/{provider}:
    delete:
      tags: [Auth]
      operationId: unlinkOwnOAuthIdentity
      summary: Unlink a Google/Facebook identity from your own account (BR-39)
      description: Removes the caller's own linked identity for the given provider, if one exists.
      x-permission: auth.oauth.unlink_own
      x-business-rules: [BR-39]
      parameters:
        - name: provider
          in: path
          required: true
          schema: { $ref: '#/components/schemas/OAuthProvider' }
      responses:
        '204': { description: Identity unlinked. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived RS256 access token from `/auth/login`, `/auth/otp/verify` or
        `/auth/refresh`. Claims: `sub` (user id), `typ` (farmer|customer|admin),
        `roles[]`, `wh` (assigned warehouse id, for Sub Warehouse Admins), `exp`.
        Role and warehouse claims are advisory for the client only — the server always
        re-resolves permissions from `rbac.json` per request.
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated UUIDv4, unique per logical operation. Required on every
        money-moving or otherwise non-repeatable POST. Replaying the same key
        with an identical body returns the original response; replaying it with a
        different body returns 409 `IDEMPOTENCY_KEY_REUSED`. Keys are retained 24 hours.
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.tohfa.in/problems/bad-request
            title: Malformed request
            status: 400
            detail: Request body is not valid JSON.
            code: BAD_REQUEST
            correlationId: 01J9F1B2C3D4E5F6G7H8J9KB
    Unauthorized:
      description: Missing, expired or invalid credentials.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.tohfa.in/problems/unauthorized
            title: Authentication required
            status: 401
            detail: The access token has expired.
            code: UNAUTHENTICATED
            correlationId: 01J9F1B2C3D4E5F6G7H8J9KC
    Forbidden:
      description: Authenticated but not permitted — RBAC or warehouse scope.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.tohfa.in/problems/forbidden
            title: Permission denied
            status: 403
            detail: Your role does not hold listing.approve.
            code: FORBIDDEN
            correlationId: 01J9F1B2C3D4E5F6G7H8J9KD
    NotFound:
      description: No such resource, or it is outside your scope.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.tohfa.in/problems/not-found
            title: Resource not found
            status: 404
            detail: No listing with id 9a8b7c6d-5e4f-4a3b-8c2d-1e0f9a8b7c6d.
            code: NOT_FOUND
            correlationId: 01J9F1B2C3D4E5F6G7H8J9KE
    Conflict:
      description: The resource is not in a state that allows this operation.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.tohfa.in/problems/invalid-state-transition
            title: Invalid state transition
            status: 409
            detail: The listing is already approved.
            code: INVALID_STATE_TRANSITION
            correlationId: 01J9F1B2C3D4E5F6G7H8J9KF
    UnprocessableEntity:
      description: Well-formed but semantically invalid — field validation or a business rule.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.tohfa.in/problems/validation-failed
            title: Validation failed
            status: 422
            detail: 2 fields failed validation.
            code: VALIDATION_FAILED
            correlationId: 01J9F1B2C3D4E5F6G7H8J9KG
            errors:
              - { field: askingPricePerKg, code: PRICE_ABOVE_CEILING, message: Must not exceed 48.00. }
              - { field: quantityKg, code: MIN, message: Must be at least 1.000 kg. }
    TooManyRequests:
      description: Rate limited.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.tohfa.in/problems/rate-limited
            title: Too many requests
            status: 429
            detail: OTP resend is limited to one per 60 seconds.
            code: RATE_LIMITED
            correlationId: 01J9F1B2C3D4E5F6G7H8J9KH
    ServerError:
      description: Unexpected server error. Quote the correlationId to support.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.tohfa.in/problems/internal-error
            title: Internal server error
            status: 500
            detail: An unexpected error occurred.
            code: INTERNAL
            correlationId: 01J9F1B2C3D4E5F6G7H8J9KJ
  schemas:
...
    Mobile:
      type: string
      pattern: '^\+91[6-9][0-9]{9}$'
      description: Indian mobile number in E.164.
      examples: ['+919876543210']
...
    CustomerRegistrationRequest:
      type: object
      required: [mobile, fullName, password]
      properties:
        mobile: { $ref: '#/components/schemas/Mobile' }
        fullName: { type: string, minLength: 2, maxLength: 120 }
        email: { type: [string, 'null'], format: email }
        password: { type: string, minLength: 10, maxLength: 128 }
        preferredLocale: { $ref: '#/components/schemas/Locale' }
        preferredWarehouseId: { type: [string, 'null'], format: uuid }
    RegistrationAccepted:
      type: object
      required: [userId, status, otpExpiresAt]
      properties:
        userId: { type: string, format: uuid }
        status: { type: string, enum: [PENDING_OTP] }
        otpExpiresAt: { type: string, format: date-time }
        resendAvailableAt: { type: string, format: date-time }
    OtpSendRequest:
      type: object
      required: [mobile, purpose]
      properties:
        mobile: { $ref: '#/components/schemas/Mobile' }
        purpose: { type: string, enum: [REGISTRATION, LOGIN, PASSWORD_RESET] }
    OtpChallenge:
      type: object
      required: [challengeId, expiresAt, attemptsRemaining]
      properties:
        challengeId: { type: string, format: uuid }
        expiresAt: { type: string, format: date-time }
        resendAvailableAt: { type: string, format: date-time, description: 60 seconds after dispatch (BR-32). }
        attemptsRemaining: { type: integer, maximum: 3, description: Lockout at 0 (BR-32). }
    OtpVerifyRequest:
      type: object
      required: [challengeId, code]
      properties:
        challengeId: { type: string, format: uuid }
        code: { type: string, pattern: '^[0-9]{6}$' }
        deviceId: { type: string, maxLength: 120 }
        platform: { type: string, enum: [ios, android, web] }
        linkToken:
          type: string
          description: |
            BR-39, optional. The `linkToken` returned by `POST /auth/oauth/{provider}`'s
            NOT_LINKED response, identifying which verified provider identity to link once this
            OTP verification succeeds. Omitted on a plain mobile+OTP verify.
    OtpVerifyResponse:
      description: |
        `TokenPair`, plus BR-39's additive/optional `oauthProfile` — present only when the
        request carried a `linkToken`, so the mobile client can prefill Step 1 of its
        registration stepper without a second API call.
      allOf:
        - $ref: '#/components/schemas/TokenPair'
        - type: object
          properties:
            oauthProfile: { $ref: '#/components/schemas/OAuthProfile' }
    LoginRequest:
      type: object
      required: [mobile, password]
      properties:
        mobile: { $ref: '#/components/schemas/Mobile' }
        password: { type: string, minLength: 1, maxLength: 128 }
        roleCode:
          type: string
          description: Required on the second call when the account holds multiple admin roles.
          enum: [SUPER_ADMIN, TOHFA_ADMIN, FARMER_ADMIN, MAIN_WH_ADMIN, SUB_WH_ADMIN]
        deviceId: { type: string, maxLength: 120 }
        platform: { type: string, enum: [ios, android, web] }
        rememberDevice: { type: boolean, default: false }
    TokenPair:
      type: object
      required: [accessToken, refreshToken, tokenType, expiresIn]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        tokenType: { type: string, enum: [Bearer] }
        expiresIn: { type: integer, description: Access token lifetime in seconds. }
        requiresRoleSelection: { type: boolean, default: false }
        availableRoles:
          type: array
          items: { $ref: '#/components/schemas/RoleAssignment' }
        user: { $ref: '#/components/schemas/CurrentUser' }
...
    CurrentUser:
      type: object
      required: [id, fullName, userType]
      properties:
        id: { type: string, format: uuid }
        fullName: { type: string }
        mobile: { $ref: '#/components/schemas/Mobile' }
        email: { type: [string, 'null'], format: email }
        userType: { type: string, enum: [farmer, customer, admin, delivery_partner] }
        preferredLocale: { $ref: '#/components/schemas/Locale' }
        roles: { type: array, items: { $ref: '#/components/schemas/RoleAssignment' } }
        permissions:
          type: array
          description: Resolved permission codes from rbac.json. Advisory for UI gating only; the server re-checks every request.
          items: { type: string }
    OAuthProvider:
      type: string
      enum: [google, facebook]
      description: The only two social-login providers TOHFA recognises (BR-39).
    OAuthLoginRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: A Google ID token, or a Facebook access token, depending on the `:provider` path segment.
        deviceId: { type: string, maxLength: 120 }
        platform: { type: string, enum: [ios, android, web] }
        roleCode:
          type: string
          description: Required on the second call when the linked account holds multiple admin roles.
          enum: [SUPER_ADMIN, TOHFA_ADMIN, FARMER_ADMIN, MAIN_WH_ADMIN, SUB_WH_ADMIN]
...
    OAuthNotLinkedResponse:
      type: object
      description: Returned by `POST /auth/oauth/{provider}` when the verified identity has no linked account yet (BR-39). No account is created; no tokens are issued.
      required: [status, profile, linkToken]
      properties:
        status: { type: string, enum: [NOT_LINKED] }
        profile: { $ref: '#/components/schemas/OAuthProfile' }
        linkToken:
          type: string
          description: Short-lived (10 minute), single-use, server-signed token. Pass to `POST /auth/otp/verify` to complete the link.
    OAuthLinkRequest:
      type: object
      required: [provider, token]
      properties:
        provider: { $ref: '#/components/schemas/OAuthProvider' }
        token:
          type: string
          description: A Google ID token, or a Facebook access token, matching `provider`.
    OAuthIdentity:
      type: object
      required: [provider, linkedAt]
      properties:
        provider: { $ref: '#/components/schemas/OAuthProvider' }
        email: { type: string, format: email, nullable: true }
        linkedAt: { type: string, format: date-time }
...